% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array_qc.R
\name{arrayqc}
\alias{arrayqc}
\title{Array Quality Control}
\usage{
arrayqc(exprs, qc_coeff = c(ks = 3, sum = 3, iqr = 3, q = 3, d = 1),
  qc_sum = 1)
}
\arguments{
\item{exprs}{An expression data.frame}

\item{qc_coeff}{A named character vector giving the coefficient for acceptable intervals}

\item{qc_sum}{Number of tests leading to array exclusion}
}
\value{
A list containing indices for columns to include and exclude
}
\description{
Array Quality Control
}
