% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_logfc.R
\name{calc_logfc}
\alias{calc_logfc}
\title{Calculate logFC}
\usage{
calc_logfc(elist)
}
\arguments{
\item{elist}{an EList with normalized expression values}
}
\value{
an EList with logFC values
}
\description{
Calculate logFC
}
