% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_nodelist.R
\name{create_nodelist}
\alias{create_nodelist}
\title{Create logFC Nodelist}
\usage{
create_nodelist(elist)
}
\arguments{
\item{elist}{An limma EList object with normalized logFC values}
}
\value{
A List with logFC values for each node of the toxicogenomic universe
}
\description{
Create logFC Nodelist
}
