% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_for_shiny.R
\name{create_shiny_data}
\alias{create_shiny_data}
\title{Create Shiny Data}
\usage{
create_shiny_data(dslist, tcta_list, ci_list)
}
\arguments{
\item{dslist}{List of logFC data (one EList per substance)}

\item{tcta_list}{fitted parameters for same substances}

\item{ci_list}{list of CI differences}
}
\value{
a list of dataframes needed by the shiny app
}
\description{
Create Shiny Data
}
