% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_for_shiny.R
\name{create_shiny_nodeplotlist}
\alias{create_shiny_nodeplotlist}
\title{Create nodeplot list for shiny app}
\usage{
create_shiny_nodeplotlist(dslist, tcta_list)
}
\arguments{
\item{dslist}{List of logFC data (one EList per substance)}

\item{tcta_list}{List of fitted parameters}
}
\value{
a list with data for nodeplots
}
\description{
Create nodeplot list for shiny app
}
