% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_for_shiny.R
\name{create_shiny_plotlist}
\alias{create_shiny_plotlist}
\title{Create Shiny Plotlist}
\usage{
create_shiny_plotlist(dslist, ci_list)
}
\arguments{
\item{dslist}{List of logFC data (one EList per substance)}

\item{ci_list}{list of CI differences}
}
\value{
a list of dataframes needed by the shiny app
}
\description{
Create Shiny Plotlist
}
