% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createkohonen.R
\name{create_tox_universe}
\alias{create_tox_universe}
\title{Create toxicogenomic universe}
\usage{
create_tox_universe(dslist, logFC_frame, dimens = 60,
  dist.fct = "manhattan", alpha = c(0.8, 0.005), includearchive = T,
  seed = 2312, output = T)
}
\arguments{
\item{dslist}{A list of ELists with smoothed logFC data to include in the learning procedure.}

\item{logFC_frame}{data.frame with additional logFC to include for map learning}

\item{dimens}{The dimension of the SOM (default: 60)}

\item{dist.fct}{The distance function to be used (default: manhattan)}

\item{alpha}{learning parameter for SOM (default: 0.8, 0.005)}

\item{includearchive}{logical, should published toxicogenomic data be included into SOM (default: T)}

\item{seed}{The seed to be used for the intial map}

\item{output}{logical, should analytical plots be given}
}
\value{
A result list containing the som_model, the final input dataset (dataset.SOM), compiled metadata, and ProbeIDs.
}
\description{
This function depends on the package "kohonen" to infer a SOM from toxicogenomic data.
}
