% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elist_median.R
\name{elist.median}
\alias{elist.median}
\title{Calculate Probe Medians of ExpressionSet}
\usage{
elist.median(data)
}
\arguments{
\item{data}{an EList}
}
\value{
an Elist
}
\description{
This function takes an EList, calculates a score from the quality flags
and removes measurements with lower scores than the minimum score for that probe. If
there are multiple measurements for the same probe the median is calculated.
}
\examples{
No example yet

}
\author{
This function is based on a function of Kristin Reiche
}
