% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_enrichment.R
\name{enrich_clusters}
\alias{enrich_clusters}
\title{Functional enrichment of universe clusters}
\usage{
enrich_clusters(cluster_table)
}
\arguments{
\item{cluster_table}{a dataframe with the columns 'ensembl' and 'clustername'}
}
\value{
returns a list with functional enrichments for ZFIN, interpro, GO, and reactome annotations
}
\description{
Functional enrichment of universe clusters
}
