% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_apical.R
\name{fit_apical}
\alias{fit_apical}
\title{Fit Apical}
\usage{
fit_apical(responsedf, plot = T, designtime = 96)
}
\arguments{
\item{responsedf}{a dataframe containing effect counts}

\item{plot}{logical, if plot should be given}

\item{designtime}{which timepoint should be taken to calculate EC values and designconcentrations}
}
\value{
a list of model fits and designconcentrations
}
\description{
Fit Apical
}
