% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_tcta.R
\name{fit_tcta}
\alias{fit_tcta}
\title{Fit regression model on toxicogenomic profile}
\usage{
fit_tcta(elist, extrema, cluster = c(TRUE, FALSE))
}
\arguments{
\item{elist}{An EList containing normalized logFC values from one exposure experiment.}

\item{extrema}{A dataframe containing extreme values for each node}

\item{cluster}{Logical, if modeling should be parallelized}
}
\value{
A dataframe with fitted model parameters for each node
}
\description{
Fit regression model on toxicogenomic profile
}
