% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcta_functions.R
\name{gauss_gauss}
\alias{gauss_gauss}
\title{Gauss-Gauss Model}
\usage{
gauss_gauss(dose, time, mconc, sconc, mu, sigma, maxGene)
}
\arguments{
\item{dose}{concentration}

\item{time}{time after exposure start}

\item{mconc}{mconc}

\item{sconc}{sconc}

\item{mu}{t_max}

\item{sigma}{simga}

\item{maxGene}{maximum logFC}
}
\value{
logFC at given time and concentration with given parameters
}
\description{
Gauss-Gauss Model
}
