% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_CI_diffs.R
\name{get_CI_diffs}
\alias{get_CI_diffs}
\title{Get Difference of Confidence Interval}
\usage{
get_CI_diffs(tcta_frame, nodelist)
}
\arguments{
\item{tcta_frame}{dataframe with fitted parameters}

\item{nodelist}{nodelist of the same experiment}
}
\value{
a list with dataframes for each node containing the CI-diffs for each treatment
}
\description{
Get Difference of Confidence Interval
}
