% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_hits.R
\name{get_hits_exons}
\alias{get_hits_exons}
\title{Get Exon Hits}
\usage{
get_hits_exons(exon_bed, mart, maxmismatch, probelengths, name = "exon",
  arraytype = c("regular", "Affy_ST"))
}
\arguments{
\item{exon_bed}{path to .bed file}

\item{mart}{biomart object}

\item{maxmismatch}{maximum number of mismatches}

\item{probelengths}{length of probes}

\item{name}{database name}

\item{arraytype}{"regular" or "Affy_ST"}
}
\value{
a data.frame containing probe-exon associations
}
\description{
Get Exon Hits
}
