% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_hits.R
\name{get_hits_genome}
\alias{get_hits_genome}
\title{Get Genome hits}
\usage{
get_hits_genome(genome_bed, mart, maxmismatch, probelengths)
}
\arguments{
\item{genome_bed}{path to .bed file}

\item{mart}{mart object}

\item{maxmismatch}{maximum number of mismatches}

\item{probelengths}{length of probes}
}
\value{
a data.frame containing probe-genome associations
}
\description{
Get Genome hits
}
