% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sig_level.R
\name{get_sig_level}
\alias{get_sig_level}
\title{Get significance level}
\usage{
get_sig_level(CIdiffs, model = c("hill-gauss", "gauss-gauss"), siglevel = 0)
}
\arguments{
\item{CIdiffs}{List of CIdiffs}

\item{model}{"hill-gauss" or "gauss-gauss"}

\item{siglevel}{level of significant (default: 0)}
}
\value{
a vector with significant levels for the map
}
\description{
Get significance level
}
