% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tcta_params.R
\name{get_tcta_params}
\alias{get_tcta_params}
\title{Get regression model parameters for one node}
\usage{
get_tcta_params(nodeframe, param_bounds)
}
\arguments{
\item{nodeframe}{A dataframe containing logFC, time, concentration of a node}

\item{param_bounds}{A list containing parameter boundaries for hill-gauss and gauss-gauss model}
}
\value{
Returns a named vector with fitted model parameters
}
\description{
Get regression model parameters for one node
}
