% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBM_annotation.R
\name{getBM_annotation}
\alias{getBM_annotation}
\title{get biomaRt annotaiton}
\usage{
getBM_annotation(values = mapFinal_reduced$ensembl_gene_id,
  filter = "ensembl_gene_id", mart)
}
\arguments{
\item{values}{IDs}

\item{filter}{Type of IDs (default: "ensembl_gene_id")}

\item{mart}{a biomaRt object}
}
\value{
a dataframe with selected biomaRt annotations
}
\description{
get biomaRt annotaiton
}
