% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{importData}
\alias{importData}
\title{Import Array Data}
\usage{
importData(targetfile, datadir, scanID = 1, output = T,
  removeOutliers = T)
}
\arguments{
\item{targetfile}{string, name of targetfile .csv, should contain columns
named FileName, sample_ID,    substance, concentration_level,    concentration_umol_l,
time_hpe,    type, comment, and    scan_ID}

\item{datadir}{string, path to raw data}

\item{scanID}{numerical, ID of scan (default: 1)}

\item{output}{logical, should function return target (default: T)}

\item{removeOutliers}{logical, should outliers be removed}
}
\value{
EListraw (limma class)
}
\description{
This function builds on the limma function read.maimages and
imports raw data from a time and concentration dependent toxicogenomic
experiment. It removes outlier arrays based on the IQR distribution and
returns an EListRaw.
}
