% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nodelist_from_logFC.R
\name{nodelist_from_logFC}
\alias{nodelist_from_logFC}
\title{Create nodelist from logFC dataframe}
\usage{
nodelist_from_logFC(logFCframe, nodeframe, substance = "substance",
  time_hpe = 1, concentration_umol_l = 1, concentration_level = "C1",
  time_hpe_factor = 1)
}
\arguments{
\item{logFCframe}{a dataframe containing the columns 'logFC' and 'ensembl_gene_id'}

\item{nodeframe}{universe nodeframe}

\item{substance}{name of substance}

\item{time_hpe}{time after exposure}

\item{concentration_umol_l}{concentration}

\item{concentration_level}{concentration level}

\item{time_hpe_factor}{time after exposure}
}
\value{
returns a nodelist
}
\description{
Create nodelist from logFC dataframe
}
