% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_batch.R
\name{normalizeBatch}
\alias{normalizeBatch}
\title{Normalize Batch}
\usage{
normalizeBatch(dslist, method = "cyclicloess", output = T)
}
\arguments{
\item{dslist}{list of EListRaw objects which should be normalized together}

\item{method}{string, normalization method (default: "cyclicloess")}

\item{output}{logical, if output plots should be given}
}
\value{
a list of EList objects containing the normalized data
}
\description{
Normalize Batch
}
