% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot3D.R
\name{plot3D_node}
\alias{plot3D_node}
\title{Plot noderesponse as surface}
\usage{
plot3D_node(nodeID, zlim, concentrations, tcta_paramframe,
  timeselect = NULL, concselect = NULL, surfacecolor = "grey",
  logconc = FALSE, plotmeasurements = FALSE, nodelist = NULL)
}
\arguments{
\item{nodelist}{}
}
\description{
Plot noderesponse as surface
}
