% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/som_plots.R
\name{plot_nodecode}
\alias{plot_nodecode}
\title{Plot Nodecode}
\usage{
plot_nodecode(nodelist, tox_universe, nodeIDs)
}
\arguments{
\item{nodelist}{A nodelist (created by toxprofileR::create_nodelist())}

\item{tox_universe}{A tox_universe (created by toxprofileR::create_universe())}

\item{nodeIDs}{Numerical, a vector of nodeIDs to be plotted}
}
\value{
Plots nodecodes and data
}
\description{
Plot Nodecode
}
