% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/som_plots.R
\name{plot_noderesponse}
\alias{plot_noderesponse}
\title{Plot noderesponse}
\usage{
plot_noderesponse(dslist, tcta_list, nodeframe, nodeID, plot3D = TRUE)
}
\arguments{
\item{dslist}{a list of ELists containing logFCs}

\item{tcta_list}{a list of parameter dataframes}

\item{nodeframe}{universe nodeframe}

\item{nodeID}{ID of node to be plotted}

\item{plot3D}{logical, should 3D plots be given}
}
\value{
returns plots of measured and fitted noderesponse
}
\description{
Plot noderesponse
}
