% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottimepoint.R
\name{plot_noderesponse_time}
\alias{plot_noderesponse_time}
\title{Plot noderesponse for selected timepoint}
\usage{
plot_noderesponse_time(dslist, tcta_list, nodeframe, nodeID, timeselect,
  concselect, mixtureranges = NULL, output = "plot", logx = F,
  plotCIs = F)
}
\arguments{
\item{dslist}{a list of ELists containing logFCs}

\item{tcta_list}{a list of parameter dataframes}

\item{nodeframe}{universe nodeframe}

\item{nodeID}{ID of node to be plotted}
}
\value{
returns plots of measured and fitted noderesponse
}
\description{
Plot noderesponse for selected timepoint
}
