% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/som_plots.R
\name{plot_portrait}
\alias{plot_portrait}
\title{Plot Tox Portrait}
\usage{
plot_portrait(nodelist, tox_universe = NULL, grid = NULL,
  tcta_paramframe = NULL, substance = NULL, time_hpe,
  concentration_level, concentration_umol_l = NULL, type = c("code",
  "median", "modeled", "parameter"), parameter = NULL,
  onlysig = c(TRUE, FALSE), siglevel = NULL, output = c("plot",
  "data"), legend = FALSE, logy = FALSE, colvec = NULL)
}
\arguments{
\item{nodelist}{A nodelist created with toxprofileR::create_nodelist()}

\item{tox_universe}{Toxicogenomic universe}

\item{grid}{grid of toxicogenomic universe}

\item{tcta_paramframe}{Optional, dataframe containing fitted parameter values for each node}

\item{substance}{Optional, substance name}

\item{time_hpe}{Time point to be plotted}

\item{concentration_level}{Concentration level to be plotted}

\item{concentration_umol_l}{concentration for prediction}

\item{type}{character string, type of response to be plotted; Possible values are  "code", "median", "modeled", "parameter"}

\item{parameter}{Optional, name of parameter to be plotted}

\item{onlysig}{logical, if reponse should be filtered to signficant values}

\item{siglevel}{vector of significant effect levels}

\item{output}{character string, if the output should be plotted ("plot") or plot data should be given as output ("data")}

\item{legend}{logical, should a legend be given out (default: F)}

\item{logy}{logical should parameter value be log-scaled}

\item{colvec}{custom vector for colorscale}
}
\value{
either a ggplot printed or ggplot data, depending on the parameter "output"
}
\description{
Plot Tox Portrait
}
