% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/som_plots.R
\name{plot_portrait_grid}
\alias{plot_portrait_grid}
\title{Plot portrait grid}
\usage{
plot_portrait_grid(nodelist, filename, tox_universe = NULL,
  grid = NULL, tcta_paramframe = NULL, substance = NULL,
  type = c("code", "median", "modeled"), parameter = NULL,
  onlysig = c(TRUE, FALSE), colvec = NULL)
}
\arguments{
\item{nodelist}{A nodelist created with toxprofileR::create_nodelist()}

\item{filename}{filename to save plot}

\item{tox_universe}{Toxicogenomic universe}

\item{grid}{grid of toxicogenomic universe}

\item{tcta_paramframe}{Optional, dataframe containing fitted parameter values for each node}

\item{substance}{Optional, substance name}

\item{type}{character string, type of response to be plotted; Possible values are  "code", "median", "modeled", "parameter"}

\item{parameter}{Optional, name of parameter to be plotted}

\item{onlysig}{logical, if reponse should be filtered to signficant values}

\item{colvec}{custom scaling for color scale}
}
\value{
Saves a fingerprint grid in the given filename
}
\description{
Plot portrait grid
}
