% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{preprocess}
\alias{preprocess}
\title{Preprocess Data}
\usage{
preprocess(elist, batchcorrect = F, batch)
}
\arguments{
\item{elist}{EList for preprocessing}

\item{batchcorrect}{logical, if batchcorrection should be applied (default: F)}

\item{batch}{factor, batch information}
}
\value{
a processed EList
}
\description{
Preprocess Data
}
