% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_raw.R
\name{read_raw_public}
\alias{read_raw_public}
\title{Read Microarray data from GEO or ArrayExpress}
\usage{
read_raw_public(datadir, rawformat = c("Agilent", "Affymetrix",
  "Affymetrix_ST"), betweenArrayNorm = "cyclicloess", metadata,
  output = T, qc_coeff = c(ks = 1.5, sum = 1.5, iqr = 1.5, q = 1.5, d =
  1), qc_sum = 2)
}
\arguments{
\item{datadir}{data directory}

\item{rawformat}{format of rawdatat (Agilen or Affymetrix)}

\item{betweenArrayNorm}{name of normalization method}

\item{metadata}{metadataframe}

\item{output}{logical if output plot should be produced}

\item{qc_coeff}{A named character vector giving the coefficient for acceptable intervals}

\item{qc_sum}{Number of tests leading to array exclusion}
}
\value{
returns a normalized Elist
}
\description{
Read Microarray data from GEO or ArrayExpress
}
