% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_outliers.R
\name{remove_outliers}
\alias{remove_outliers}
\title{Remove outliers}
\usage{
remove_outliers(nodedf, plots = F)
}
\arguments{
\item{nodedf}{A data.frame containing a "logFC" column}

\item{plots}{logical, should plots with outliers be shown?}
}
\value{
The input data.frame with outlying measurements removed
}
\description{
Remove outliers
}
