% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_for_shiny.R
\name{save_shiny_data}
\alias{save_shiny_data}
\title{Save Shiny Data}
\usage{
save_shiny_data(dslist, ci_list, tcta_list, martversion, filename)
}
\arguments{
\item{dslist}{List of logFC data (one EList per substance)}

\item{ci_list}{list of CI differences}

\item{tcta_list}{List of fitted parameters}

\item{martversion}{martversion}

\item{filename}{path and filename to be stored}
}
\value{
saves an RDS file with data for shiny
}
\description{
Save Shiny Data
}
