% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spline_fit.R
\name{spline_fit}
\alias{spline_fit}
\title{Fit 3-D spline to time and concentration dependent data}
\usage{
spline_fit(elist)
}
\arguments{
\item{elist}{an EList with logFC data}
}
\value{
an EList with smoothed logFC data
}
\description{
Fit 3-D spline to time and concentration dependent data
}
