% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array_qc.R
\name{arrayqc}
\alias{arrayqc}
\title{Array Quality Control}
\usage{
arrayqc(exprs)
}
\arguments{
\item{exprs}{An expression data.frame}
}
\value{
A list containing indices for columns to include and exclude
}
\description{
Array Quality Control
}
