% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateFC_public.R
\name{calc_logfc_public}
\alias{calc_logfc_public}
\title{Calculate logFC of public microarray data}
\usage{
calc_logfc_public(data, comparison)
}
\arguments{
\item{data}{EList}

\item{comparison}{data.frame describing the comparison for logFC}
}
\value{
dataframe of logFCs
}
\description{
Calculate logFC of public microarray data
}
