% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_apical.R
\name{fit_apical}
\alias{fit_apical}
\title{Fit Apical}
\usage{
fit_apical(responsedf, plot = T, designtime = 96,
  startlogit = list(t1 = -1, t2 = -1), startweibull = list(t1 = 10, t2
  = 10), startgl = list(t1 = 10, t2 = 10, t3 = 0.01))
}
\arguments{
\item{responsedf}{a dataframe containing effect counts}

\item{plot}{logical, if plot should be given}

\item{designtime}{which timepoint should be taken to calculate EC values and designconcentrations}

\item{startlogit}{list of startparameters for logit}

\item{startweibull}{list of startparameters for weibull}

\item{startgl}{list of startparameters for gl}
}
\value{
a list of model fits and designconcentrations
}
\description{
Fit Apical
}
