% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{importData}
\alias{importData}
\title{Import Array Data}
\usage{
importData(targetfile, datadir, scanID = 1, output = T,
  removeOutliers = T, qc_coeff = c(ks = 1.5, sum = 1.5, iqr = 1.5, q =
  1.5, d = 1), qc_sum = 2)
}
\arguments{
\item{targetfile}{string, name of targetfile .csv, should contain columns
named FileName, sample_ID,    substance, concentration_level,    concentration_umol_l,
time_hpe,    type, comment, and    scan_ID}

\item{datadir}{string, path to raw data}

\item{scanID}{numerical, ID of scan (default: 1)}

\item{output}{logical, should function return target (default: T)}

\item{removeOutliers}{logical, should outliers be removed}

\item{qc_coeff}{A named character vector giving the coefficient for acceptable intervals}

\item{qc_sum}{Number of tests leading to array exclusion}
}
\value{
EListraw (limma class)
}
\description{
This function builds on the limma function read.maimages and
imports raw data from a time and concentration dependent toxicogenomic
experiment. It removes outlier arrays based on the KS.test, and outlier test based on sum and 75% quantile. The function
returns an EListRaw.
}
