% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/som_plots.R
\name{plot_portrait}
\alias{plot_portrait}
\title{Plot Tox Portrait}
\usage{
plot_portrait(nodelist, tox_universe, tcta_paramframe = NULL,
  substance = NULL, time_hpe, concentration_level, type = c("code",
  "median", "modeled", "parameter"), parameter = NULL,
  onlysig = c(TRUE, FALSE), output = c("plot", "data"),
  legend = FALSE)
}
\arguments{
\item{nodelist}{A nodelist created with toxprofileR::create_nodelist()}

\item{tox_universe}{Toxicogenomic universe}

\item{tcta_paramframe}{Optional, dataframe containing fitted parameter values for each node}

\item{substance}{Optional, substance name}

\item{time_hpe}{Time point to be plotted}

\item{concentration_level}{Concentration level to be plotted}

\item{type}{character string, type of response to be plotted; Possible values are  "code", "median", "modeled", "parameter"}

\item{parameter}{Optional, name of parameter to be plotted}

\item{onlysig}{logical, if reponse should be filtered to signficant values}

\item{output}{character string, if the output should be plotted ("plot") or plot data should be given as output ("data")}

\item{legend}{logical, should a legend be given out (default: F)}
}
\value{
either a ggplot printed or ggplot data, depending on the parameter "output"
}
\description{
Plot Tox Portrait
}
