% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/som_plots.R
\name{plot_portrait_grid}
\alias{plot_portrait_grid}
\title{Plot portrait grid}
\usage{
plot_portrait_grid(nodelist, tox_universe, tcta_paramframe = NULL,
  substance = NULL, type = c("code", "median", "modeled", "parameter"),
  parameter = NULL, onlysig = c(TRUE, FALSE), output = c("plot",
  "data"))
}
\arguments{
\item{nodelist}{A nodelist created with toxprofileR::create_nodelist()}

\item{tox_universe}{Toxicogenomic universe}

\item{tcta_paramframe}{Optional, dataframe containing fitted parameter values for each node}

\item{substance}{Optional, substance name}

\item{type}{character string, type of response to be plotted; Possible values are  "code", "median", "modeled", "parameter"}

\item{parameter}{Optional, name of parameter to be plotted}

\item{onlysig}{logical, if reponse should be filtered to signficant values}

\item{output}{character string, if the output should be plotted ("plot") or plot data should be given as output ("data")}
}
\value{
Returns a grid plot of time and concentration portraits
}
\description{
Plot portrait grid
}
