% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_raw.R
\name{read_raw_public}
\alias{read_raw_public}
\title{Read Microarray data from GEO or ArrayExpress}
\usage{
read_raw_public(datadir, rawformat = c("Agilent", "Affymetrix",
  "Affymetrix_ST"), betweenArrayNorm = "cyclicloess", metadata)
}
\arguments{
\item{datadir}{data directory}

\item{rawformat}{format of rawdatat (Agilen or Affymetrix)}

\item{betweenArrayNorm}{name of normalization method}

\item{metadata}{metadataframe}
}
\value{
returns a normalized Elist
}
\description{
Read Microarray data from GEO or ArrayExpress
}
